refactor: 디렉토리 이름 및 구조 변경 - 프로젝트 구조를 개선하기 위해 디렉토리 이름 변경 및 파일 구조 재정비
